backend/requirements.txt
asgiref==3.11.0
channels==4.0.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
channels-redis==4.2.0
redis==5.0.8
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0